Revert breaking changes from 0.7.0
In release 0.7.0 we introduced some changes to facilitate integration
between Haystack and Adaptive Alerting. Unfortunately these changes broke
serialization and deserialization for existing Adaptive Alerting clients
and we need to roll them back.

The changes in 0.7.0 were anyway a temporary hack to support the aforesaid
integration. We will have to think of a more permanent way to handle this.